standardize day3 p2

+++ src/day3.h
@@ -5,6 +5,6 @@
 
 int day_driver(FILE *, FILE *);
 int day3_p1(FILE *);
-long int day3_p2(FILE *);
+long long day3_p2(FILE *);
 
 #endif
